+++ lib/onlyoffice_webdriver_wrapper/webdriver.rb
@@ -169,10 +169,6 @@ module OnlyofficeWebdriverWrapper
       WebDriver.host_name_by_full_url(get_url)
     end
 
-    def remove_event(event_name)
-      execute_javascript("jQuery(document).unbind('#{event_name}');")
-    end
-
     def remove_class_by_jquery(selector, class_name)
       execute_javascript("$(#{dom_element_by_xpath(selector)}).removeClass('#{class_name}');")
     end